Crate gba2k

A GBA development project.

Safety

All of the crate’s safety calculations assume that you’re using our provided linker script, and using the thumbv4t-none-eabi target.

On any other target you can still render the crate’s docs, but you generally won’t be able to compile the crate. Even if you can, all MMIO and inline assembly would be incorrect to run.

Modules

bios

The BIOS of the GBA provides some code that’s built-in to the device.

interrupts

Module for interacting with the GBA’s hardware interrupt system.

keys

Allows reading the GBA’s key inputs.

rt0

The rt0 is the assembly runtime that runs before your Rust code.

video

Module for utilizing the GBA’s screen.

Macros

read_lr

Inline assembly to read the link register value.

read_pc

Inline assembly to read the program counter value.

read_sp

Inline assembly to read the stack pointer value.