Struct gba2k::interrupts::GbaCell

#[repr(transparent)]
pub struct GbaCell<T: GbaCellSafe>(_);

A GbaCell holds a value that’s accessed in a single machine instruction.

Use of the GbaCell type allows for safe mutable global data that can be accessed by both the main program as well as the interrupt handler.

Implementations

impl<T: GbaCellSafe> GbaCell<T>

pub const fn new(t: T) -> Self

Makes a new cell.

pub fn read(&self) -> T

Reads the current value.

pub fn write(&self, t: T)

Writes a new value.

pub fn get(&self) -> *mut T

Gets the raw pointer to the value.

Safety

A GbaCell is an [UnsafeCell] internally, and so this raw pointer must follow all the usuall “pointer into an UnsafeCell” rules.