Struct gba2k::keys::KeyInput

#[repr(transparent)]
pub struct KeyInput(_);

Key state data.

Each method uses true for pressed and false for released. Internally, the actual bits from the hardware are 0 when pressed and 1 when released, but all the methods will automatically compensate for this. The only time it should come up is if you’re converting to/from a u16 value.

Implementations

impl KeyInput

pub const fn new() -> Self

pub const fn a(self) -> bool

pub const fn with_a(self, val: bool) -> Self

pub const fn b(self) -> bool

pub const fn with_b(self, val: bool) -> Self

pub const fn select(self) -> bool

pub const fn with_select(self, val: bool) -> Self

pub const fn start(self) -> bool

pub const fn with_start(self, val: bool) -> Self

pub const fn right(self) -> bool

pub const fn with_right(self, val: bool) -> Self

pub const fn left(self) -> bool

pub const fn with_left(self, val: bool) -> Self

pub const fn up(self) -> bool

pub const fn with_up(self, val: bool) -> Self

pub const fn down(self) -> bool

pub const fn with_down(self, val: bool) -> Self

pub const fn r(self) -> bool

pub const fn with_r(self, val: bool) -> Self

pub const fn l(self) -> bool

pub const fn with_l(self, val: bool) -> Self

pub const fn changes_since(self, previous: KeyInput) -> KeyChanges

Determines what keys have changed since a previous KeyInput.