Module gba2k::rt0

The rt0 is the assembly runtime that runs before your Rust code.

This module also serves as the home of any other handwritten assembly files that the crate adds over time.

Functions

__aeabi_idiv⚠

Performs i32 / i32 division.

__aeabi_idivmod⚠

Performs [i32 / i32, i32 % i32], returning the data packed in a u64.

__aeabi_memclr⚠

Sets n bytes in dest to 0.

__aeabi_memclr4⚠

Sets n bytes in dest to 0.

__aeabi_memclr8⚠

Sets n bytes in dest to 0.

__aeabi_memcpy⚠

Copies n bytes to dest from src.

__aeabi_memcpy4⚠

Copies n bytes to dest from src.

__aeabi_memcpy8⚠

Copies n bytes to dest from src.

__aeabi_memmove⚠

Copies n bytes to dest from src (overlap allowed).

__aeabi_memmove4⚠

Copies n bytes to dest from src (overlap allowed).

__aeabi_memmove8⚠

Copies n bytes to dest from src (overlap allowed).

__aeabi_memset⚠

Sets n bytes in dest to val as u8.

__aeabi_memset4⚠

Sets n bytes in dest to val as u8.

__aeabi_memset8⚠

Sets n bytes in dest to val as u8.

__aeabi_uidiv⚠

Performs u32 / u32 division.

__aeabi_uidivmod⚠

Performs [u32 / u32, u32 % u32], returning the data packed in a u64.

__aeabi_uread4⚠

Safely performs an unaligned read of the u32 address given.

__aeabi_uread8⚠

Safely performs an unaligned read of the u64 address given.

__aeabi_uwrite4⚠

Safely performs an unaligned write of the u32 address given.

__aeabi_uwrite8⚠

Safely performs an unaligned write of the u64 address given.

gba_memcpy_sram⚠

Copies n bytes to dest from src.

memcpy⚠

Copies n bytes to dest from src.

memmove⚠

Copies n bytes to dest from src (overlap allowed).

memset⚠

Sets n bytes in dest to val as u8.

set_rust_irq_handler

Sets the rust function to run when a hardware interrupt occurs.

text_single_swp⚠

A swp instruction, stored in the .text section.

text_single_swpb⚠

A swpb instruction, stored in the .text section.