Function gba2k::rt0::__aeabi_memcpy

pub unsafe extern "C" fn __aeabi_memcpy(
    dest: *mut u8, 
    src: *mut u8, 
    count: usize
)

Provides ARM-style memory copy.

This works just like memcpy from libc except that there’s no return value. This allows a small saving on stack space, as well as saving a few cycles required to push/pop two words.

Prefer this function over memcpy whenever you don’t need the return value.

If you know that both your pointers are aligned to 4, prefer __aeabi_memcpy4.

Safety

The src and dest regions must not overlap.

src must be valid to read for count bytes.

dest must be valid to write for count bytes.