Function gba2k::rt0::set_rust_irq_handler

pub fn set_rust_irq_handler(opt_f: Option<extern "C" fn(_: IrqBits)>)

Sets the rust function to run when a hardware interrupt occurs.

When a hardware interrupt happens, control transfers to the BIOS, which calls the rt0 interrupt handler, which will acknowledge the interrupt(s) and then call your handler with the IrqBits of what interrupt(s) just occurred.

If you set the handler to None the rt0 handler will still acknowledge any interrupt, but will just skip calling back into rust.