Module gba2k::video

Module for utilizing the GBA’s screen.

Video Modes

The GBA has 6 video modes. They’re just referred to by their index, 0 through 5.

Currently the crate only provides direct support for Mode 3. Support for other modes is planned.

Modules

mode3

Video Mode 3: A single 240x160 bitmap.

Structs

Color

The GBA has RGB color with 5 bits per channel (“RGB555”).

DisplayControl

The display control setting is the top level control of the screen.

DisplayStatus

The display status lets you get info about the display and configure display interrupts.

Enums

VideoMode

Constants

BACKDROP_COLOR

The color of the screen’s “backdrop”.

BG_PALETTE

The background palette entries.

DISPCNT

“Display Control”

DISPSTAT

“Display Status”

OBJ_PALETTE

The object palette entries.

VCOUNT

Returns the current scanline that the PPU is drawing (0 through 227).

Functions

bg_palbank

Gets palbank x of the background palette.

obj_palbank

Gets palbank x of the object palette.

Type Definitions

ColorAddress

A volatile address for a Color value.

PalBank

Alias for an individual palbank (16 color group) within the palette.