Struct gba2k::video::DisplayControl

#[repr(transparent)]
pub struct DisplayControl(_);

The display control setting is the top level control of the screen.

Primarily, the display control sets the video mode. Also, it lets you set which layers are displayed or not. Finally, it has a few other bits that control a few random other parts of the video system.

Implementations

impl DisplayControl

pub const fn new() -> Self

pub const fn video_mode(self) -> VideoMode

pub const fn with_video_mode(self, val: VideoMode) -> Self

pub const fn frame_select(self) -> bool

pub const fn with_frame_select(self, val: bool) -> Self

pub const fn hblank_oam_free(self) -> bool

pub const fn with_hblank_oam_free(self, val: bool) -> Self

pub const fn obj_vram_1d(self) -> bool

pub const fn with_obj_vram_1d(self, val: bool) -> Self

pub const fn forced_blank(self) -> bool

pub const fn with_forced_blank(self, val: bool) -> Self

pub const fn display_bg0(self) -> bool

pub const fn with_display_bg0(self, val: bool) -> Self

pub const fn display_bg1(self) -> bool

pub const fn with_display_bg1(self, val: bool) -> Self

pub const fn display_bg2(self) -> bool

pub const fn with_display_bg2(self, val: bool) -> Self

pub const fn display_bg3(self) -> bool

pub const fn with_display_bg3(self, val: bool) -> Self

pub const fn display_obj(self) -> bool

pub const fn with_display_obj(self, val: bool) -> Self

pub const fn display_win0(self) -> bool

pub const fn with_display_win0(self, val: bool) -> Self

pub const fn display_win1(self) -> bool

pub const fn with_display_win1(self, val: bool) -> Self

pub const fn display_obj_win(self) -> bool

pub const fn with_display_obj_win(self, val: bool) -> Self