Module gba2k::rt0

The rt0 is the assembly runtime that runs before your Rust code.

This module also serves as the home of any other handwritten assembly files that the crate adds over time.

Functions

__aeabi_memcpy⚠

Provides ARM-style memory copy.

__aeabi_memcpy4⚠

Works like __aeabi_memcpy but optimized for pointers aligned to 4.

__aeabi_memcpy8⚠

Works like __aeabi_memcpy but optimized for pointers aligned to 8.

__aeabi_memmove⚠

Like __aeabi_memcpy, but the regions can overlap.

__aeabi_memmove4⚠

Like __aeabi_memcpy4, but the regions can overlap.

__aeabi_memmove8⚠

Like __aeabi_memcpy8, but the regions can overlap.

memcpy⚠

Provides libc-style memory copy.

memcpy_sram⚠

A memory copy operation guaranteed to work correctly with SRAM.

memmove⚠

Like memcpy, but the regions can overlap.

set_rust_irq_handler

Sets the rust function to run when a hardware interrupt occurs.

text_single_swp⚠

A swp instruction, stored in the .text section.

text_single_swpb⚠

A swpb instruction, stored in the .text section.